Refactor the make targets
Allows to rerun fixtures only if the fixture definition changed.
Additionally the install target got simpler and a clear tree of dependencies.
To enable those improvements a new temporary directory was needed.
Marker files allow tracking of one time changes or changes relativ to other files.
The implementation is a step toward issue #8.