Add first ansible task to automate deployment
With this commit you can install php, mysql and java via ansible.
It also provides tasks to download and unpack keycloak.
Which version of keycloak is determined by variables within the playbook.
This is depended on the structure of the jboss download server and may break.
The current implementation tries to avoid more complex logic within the playbook.
Additionally all tasks run idempotent and nothing should be skipped.