Add inventories target
All possible inventories for this project are simple.
They will usually contain only one ip/machine and
to prevent any fraud, no ip's are saved in the repository anymore.
Therefore you now need to generate the inventories from the makevars.
You can do that manually with the inventories target or
it will happen automatically, if you run the deploy target.