Add ansible_user to generated inventories
You can now provide a ansible_user within the makevars file.
For the variable name within the makefile, please refer to the variable defaults section.
The username will be written out in your inventories.
This is useful, when your development machine user is not the same as the deploy target user.
This solves issue #30.