Fix false positiv test caused by webpack
Running the webpack dev-server brings long waiting times.
If you run phpunit shortly after an initial install,
you can run into failed tests because of a missing asset entrypoint.
This is of course not broken code and is fixed by precompiling the assets,
before a test or a local instance runs.