Fix debug-toolbar deprecations
The security config syntax changed.
The Kernel.php file was updated according to the upstream changes of the latest skeleton.
Additionally index.php was updated without effects.
Unfortunately non of the phpUnit bridge deprecations could be fixed.
Their origin is not clear and for monolog (7 of 8 deprecations) a fix comes from the bundle.