Refactor Review feature into its own Controller
These change changes decouple the review page from the actual entities.
Currently only filled with dummy data.
The embedded controllers run fast enough for up to 30 calls without
caching.
If the number of displayed metdata exceeds this number a different
solution should be researched.
Caching with esi should be done in the future.